separate out Handler module
