+++ daemon/ghc-specter-daemon.cabal
@@ -123,6 +123,7 @@ executable ghc-specter-daemon-terminal
 executable ghc-specter-daemon
   main-is: Main.hs
   other-modules:
+    Handler
     ImGuiMain
     Util.Color
     Util.GUI
